refactor: ceramic module + api functions arguments
